Update to react-native 0.34.1
