+++ package.json
@@ -30,8 +30,8 @@
   "dependencies": {
     "alt": "^0.18.4",
     "moment": "^2.13.0",
-    "react": "15.0.2",
-    "react-native": "0.26.3",
+    "react": "15.3.2",
+    "react-native": "0.34.1",
     "react-native-navbar": "^1.5.0",
     "react-native-router-flux": "^3.26.24",
     "react-native-simple-store": "^1.0.1",
@@ -39,8 +39,10 @@
     "rn-viewpager": "^1.1.2"
   },
   "devDependencies": {
-    "babel-eslint": "^6.0.4",
-    "eslint": "^2.11.1",
-    "eslint-plugin-react": "^5.1.1"
+    "eslint": "^3.6.1",
+    "eslint-config-airbnb": "^12.0.0",
+    "eslint-plugin-import": "^2.0.0",
+    "eslint-plugin-jsx-a11y": "^2.2.2",
+    "eslint-plugin-react": "^6.3.0"
   }
 }
